+++ src/vector/nalgebra_svector.rs
@@ -6,6 +6,8 @@ use nalgebra::{DMatrix, SMatrix, SVector};
 
 #[cfg(feature = "nalgebra")]
 impl<const N: usize, S: Scalar> Vector<S> for SVector<S, N> {
+    type Vectorf64 = SVector<f64, N>;
+
     type MatrixNxN = SMatrix<S, N, N>;
 
     type MatrixMxN<const M: usize> = SMatrix<S, M, N>;